+++ test/integration/NonExposedModule/Exposed.hs
@@ -1,0 +1,3 @@
+module NonExposedModule.Exposed (foo) where
+
+import NonExposedModule.NoImplicitImport (foo)